connect mocked keystones using EPHash
