+++ service/testutil/mock.go
@@ -3,6 +3,7 @@ package testutil
 import (
 	"cmp"
 	"context"
+	"crypto/sha256"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -20,6 +21,41 @@ import (
 	"github.com/hemilabs/heminetwork/api/tbcapi"
 	"github.com/hemilabs/heminetwork/hemi"
 )
+
+func MakeSharedKeystones(n int) (map[chainhash.Hash]*hemi.L2KeystoneAbrev, []hemi.L2Keystone) {
+	kssList := make([]hemi.L2Keystone, n)
+	kssMap := make(map[chainhash.Hash]*hemi.L2KeystoneAbrev, 0)
+
+	prevKeystone := &hemi.L2Keystone{
+		Version:       1,
+		L1BlockNumber: 0xbadc0ffe,
+		EPHash:        digest256([]byte{0xde, 0xad, 0xbe, 0xef}),
+	}
+	for ci := range n {
+		x := uint8(ci)
+		l2Keystone := hemi.L2Keystone{
+			Version:            1,
+			L1BlockNumber:      prevKeystone.L1BlockNumber + 1,
+			L2BlockNumber:      uint32(ci+1) * 25,
+			ParentEPHash:       digest256([]byte{x}),
+			PrevKeystoneEPHash: prevKeystone.EPHash,
+			StateRoot:          digest256([]byte{x, x, x}),
+			EPHash:             digest256([]byte{x, x, x, x}),
+		}
+
+		abrevKss := hemi.L2KeystoneAbbreviate(l2Keystone)
+		kssMap[*abrevKss.Hash()] = abrevKss
+		kssList[ci] = l2Keystone
+		prevKeystone = &l2Keystone
+	}
+
+	return kssMap, kssList
+}
+
+func digest256(x []byte) []byte {
+	xx := sha256.Sum256(x)
+	return xx[:]
+}
 
 // Opgeth RPC Messages structs
